Day 10
Rewriting section 5, and 6 for Lecture 3 insertion sort. This is slow in the beginning to build up all the prereqs we will need but will get faster as we gain this experience. Spending 2-3 days on one chapter will become a 10 minute chapter review. One lecture a week will become one lecture a day.